Rename thread.py to threads.py
There are multiple threads.
All the other files follow a plural naming scheme.
